requirements.txt
requests
feedparser
psycopg2-binary
python-dotenv